# ch_4/api-docs.json
{
  "info": {
    "description": "Banking System",
    "title": "Swagger API",
    "version": "1.0.0"
  },
  "host": "localhost:undefined",
  "basePath": "/api/v1",
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User's data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "name": "user",
            "in": "body",
            "description": "User's name",
            "required": false,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User's email",
            "required": false,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User's password",
            "required": false,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Profile information",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          {
            "name": "user",
            "in": "body",
            "description": "Type of identity (e.g., KTP, KK)",
            "required": false,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Unique identity number",
            "required": false,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User's address",
            "required": false,
            "type": "string"
          }
        ],
        "description": "",
        "tags": [
          "Users"
        ],
        "responses": {
          "201": {
            "description": "Successfully created user"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new user and a new profile"
      },
      "get": {
        "parameters": [],
        "description": "",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all users"
      }
    },
    "/api/v1/users/:userId": {
      "get": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer"
          }
        ],
        "description": "",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User's detail with profile"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get user detail by ID"
      },
      "delete": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer"
          }
        ],
        "description": "",
        "tags": [
          "Users"
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete a user by ID"
      }
    }
  },
...
  "securityDefinitions": {

  },
...
}